Initialized Controllers + Tests + Decklists
Stubs for controllers
