+++ DeckLists/Hero/TheStrangerDeckList.json
@@ -448,7 +448,7 @@
             "flavorReference": "The Stranger, Strange Roads #187"
         }
     ],
-    "notPromoCards": [
+    "promoCards": [
         {
             "identifier": "TheStrangerCharacter",
             "promoIdentifier": "RunecarvedTheStrangerCharacter",
@@ -583,35 +583,33 @@
             "promoIdentifier": "CornTheStrangerCharacter",
             "count": 1,
             "title": "The Stranger",
-            "promoTitle": "Stranger in the Corn",
+            "promoTitle": "The Stranger in the Corn",
             "shortPromoTitle": "Corn Stranger",
-            "body": "",
-            "backgroundColor": "",
+            "body": "Fields of Terror",
+            "backgroundColor": "ECA547",
             "foilBackgroundColor": "FEFEFE",
-            "hitpointColor": "",
+            "hitpointColor": "ECA547",
             "character": true,
             "powers": [
-                ""
+                "{TheStranger} or a target next to a Rune deals 1 target 2 psychic damage."
             ],
             "icons": [
                 "HasPower",
-                "Discard",
-                "DealDamageInfernal"
+                "DealDamagePsychic"
             ],
-            "hitpoints": 1,
+            "hitpoints": 27,
             "nemesisIdentifiers": [
                 "TheStranger"
             ],
             "incapacitatedAbilities": [
-                "",
-                "",
-                ""
+                "One player may draw a card now.",
+                "One hero may use a power now.",
+                "One target deals itself 1 irreducible toxic damage."
             ],
             "flippedIcons": [
                 "DrawCardNow",
-                "DealDamageInfernal",
-                "DestroyEnvironment",
-                "PlayCardNow"
+                "UsePowerNow",
+                "DealDamageToxic"
             ],
             "openingLines": {
                 "default": "Flee in terror at my visage!",
